refactor : 과제 코드 추가 및 코드리뷰 내용 반영
