+++ app/src/main/java/com/codesoom/assignment/models/TaskList.java
@@ -4,13 +4,13 @@ import com.codesoom.assignment.config.TaskNotFoundException;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Optional;
 
 public class TaskList {
     private final List<Task> taskList = new ArrayList<>();
+    private int newId = 0;
 
     public void add(Task task) {
-        setTaskId(task);
+        task.setId(generatedId());
         taskList.add(task);
     }
 
@@ -18,14 +18,8 @@ public class TaskList {
         return this.taskList;
     }
 
-    private void setTaskId(Task task) {
-        if (size() == 0) {
-            task.setId(1);
-            return;
-        }
-
-        Task laskTask = this.taskList.get(size() - 1);
-        task.setId(laskTask.getId() + 1);
+    private int generatedId() {
+        return newId += 1;
     }
 
     public int size() {
@@ -33,15 +27,10 @@ public class TaskList {
     }
 
     public Task get(int idx) throws TaskNotFoundException {
-        Optional<Task> taskOptional = taskList.stream()
+        return taskList.stream()
                 .filter(task -> task.isTaskId(idx))
-                .findFirst();
-
-        if (taskOptional.isEmpty()) {
-            throw new TaskNotFoundException("taskEmpty");
-        }
-
-        return taskOptional.get();
+                .findFirst()
+                .orElseThrow(() -> new TaskNotFoundException("tasks not found"));
     }
 
     public boolean delete(int requestTaskId) throws TaskNotFoundException {
